// src/api/user/api.ts
export interface AuthProvider {
    /**
     * The authentication provider for the user.
     * @type {string}
     * @memberof AuthProvider
     */
    'authProvider': AuthProviderAuthProviderEnum;
    /**
     * The ID token from the authentication provider (e.g., Google or Microsoft).
     * @type {string}
     * @memberof AuthProvider
     */
    'authProviderIdToken'?: string;
}

export const AuthProviderAuthProviderEnum = {
    Google: 'google',
    Microsoft: 'microsoft',
    Facebook: 'facebook',
    Email: 'email'
} as const;

export type AuthProviderAuthProviderEnum = typeof AuthProviderAuthProviderEnum[keyof typeof AuthProviderAuthProviderEnum];
// ...
export interface LoginRequest {
    /**
     * 
     * @type {AuthProvider}
     * @memberof LoginRequest
     */
    'provider': AuthProvider;
    /**
     * 
     * @type {UserCredentials}
     * @memberof LoginRequest
     */
    'credentials': UserCredentials;
}
// ...
export interface RegisterRequest {
    /**
     * 
     * @type {AuthProvider}
     * @memberof RegisterRequest
     */
    'provider': AuthProvider;
    /**
     * 
     * @type {UserRegistrationDetails}
     * @memberof RegisterRequest
     */
    'user': UserRegistrationDetails;
}
/**
 * 
 * @export
 * @interface ResendOtpRequest
 */
export interface ResendOtpRequest {
    /**
     * The user\'s email address to which the OTP will be resent.
     * @type {string}
     * @memberof ResendOtpRequest
     */
    'email': string;
}
/**
 * 
 * @export
 * @interface SendEmailOtpRequest
 */
export interface SendEmailOtpRequest {
    /**
     * The user\'s email to which the OTP will be sent.
     * @type {string}
     * @memberof SendEmailOtpRequest
     */
    'email': string;
}
/**
 * 
 * @export
 * @interface Tokens
 */
export interface Tokens {
    /**
     * The JWT access token for authenticating subsequent requests.
     * @type {string}
     * @memberof Tokens
     */
    'accessToken'?: string;
    /**
     * The expiration date and time of the access token in ISO 8601 format.
     * @type {string}
     * @memberof Tokens
     */
    'accessTokenExpiresAt'?: string;
    /**
     * The refresh token for obtaining new access tokens.
     * @type {string}
     * @memberof Tokens
     */
    'refreshToken'?: string;
    /**
     * The expiration date and time of the refresh token in ISO 8601 format.
     * @type {string}
     * @memberof Tokens
     */
    'refreshTokenExpiresAt'?: string;
}
// ...
export interface UserAuthenticationResponseData {
    data(data: any): unknown;
    access_token: any;
    /**
     * 
     * @type {UserResponse}
     * @memberof UserAuthenticationResponseData
     */
    'user'?: UserResponse;
    /**
     * 
     * @type {Tokens}
     * @memberof UserAuthenticationResponseData
     */
    'tokens'?: Tokens;
}
/**
 * 
 * @export
 * @interface UserCredentials
 */
export interface UserCredentials {
    /**
     * The user\'s email address.
     * @type {string}
     * @memberof UserCredentials
     */
    'email': string;
    /**
     * The user\'s password.
     * @type {string}
     * @memberof UserCredentials
     */
    'password'?: string;
}
// ...
export interface UserResponse {
    /**
     * The unique ID of the registered user.
     * @type {string}
     * @memberof UserResponse
     */
    'id'?: string;
    /**
     * business data
     * @type {string}
     * @memberof UserResponse
     */
    'businessName'?: string;
    /**
     * business address
     * @type {string}
     * @memberof UserResponse
     */
    'businessAddress'?: string;
    /**
     * The user\'s email address.
     * @type {string}
     * @memberof UserResponse
     */
    'email'?: string;
    /**
     * URL to the user\'s profile picture.
     * @type {string}
     * @memberof UserResponse
     */
    'picture'?: string;
    /**
     * User Tax number
     * @type {string}
     * @memberof UserResponse
     */
    'taxNumber'?: string;
    /**
     * The user\'s phone number in international format.
     * @type {string}
     * @memberof UserResponse
     */
    'phoneNumber'?: string;
    /**
     * A country where business is hosted.
     * @type {string}
     * @memberof UserResponse
     */
    'country'?: string;
    /**
     * A city where business is hosted.
     * @type {string}
     * @memberof UserResponse
     */
    'city'?: string;
    /**
     * The postal code or ZIP code of the user\'s business location.
     * @type {string}
     * @memberof UserResponse
     */
    'zipCode'?: string;
    /**
     * URL to the user\'s website.
     * @type {string}
     * @memberof UserResponse
     */
    'webSite'?: string;
    /**
     * A country state where business is hosted.
     * @type {string}
     * @memberof UserResponse
     */
    'state'?: string;
}
// ...
export interface UserUpdateDetails {
    /**
     * 
     * @type {string}
     * @memberof UserUpdateDetails
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserUpdateDetails
     */
    'phone'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserUpdateDetails
     */
    'country'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserUpdateDetails
     */
    'companyName'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserUpdateDetails
     */
    'companyAddress'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserUpdateDetails
     */
    'companyTaxNumber'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserUpdateDetails
     */
    'zipCode'?: string;
    /**
     * 
     * @type {string}
     * @memberof UserUpdateDetails
     */
    'site'?: string;
}
/**
 * 
 * @export
 * @interface VerifyEmailOtpRequest
 */
export interface VerifyEmailOtpRequest {
    /**
     * The user\'s email address.
     * @type {string}
     * @memberof VerifyEmailOtpRequest
     */
    'email': string;
    /**
     * The OTP code sent to the user\'s email.
     * @type {string}
     * @memberof VerifyEmailOtpRequest
     */
    'otp': string;
}
// ...
export interface UserApiLoginRequest {
    /**
     * 
     * @type {LoginRequest}
     * @memberof UserApiLogin
     */
    readonly loginRequest: LoginRequest
}

/**
 * Request parameters for registerUser operation in UserApi.
 * @export
 * @interface UserApiRegisterUserRequest
 */
export interface UserApiRegisterUserRequest {
    /**
     * 
     * @type {RegisterRequest}
     * @memberof UserApiRegisterUser
     */
    readonly registerRequest: RegisterRequest
}

/**
 * Request parameters for resendEmailOtp operation in UserApi.
 * @export
 * @interface UserApiResendEmailOtpRequest
 */
export interface UserApiResendEmailOtpRequest {
    /**
     * 
     * @type {ResendOtpRequest}
     * @memberof UserApiResendEmailOtp
     */
    readonly resendOtpRequest: ResendOtpRequest
}

/**
 * Request parameters for sendEmailOtp operation in UserApi.
 * @export
 * @interface UserApiSendEmailOtpRequest
 */
export interface UserApiSendEmailOtpRequest {
    /**
     * 
     * @type {SendEmailOtpRequest}
     * @memberof UserApiSendEmailOtp
     */
    readonly sendEmailOtpRequest: SendEmailOtpRequest
}

/**
 * Request parameters for updateCurrentUser operation in UserApi.
 * @export
 * @interface UserApiUpdateCurrentUserRequest
 */
export interface UserApiUpdateCurrentUserRequest {
    /**
     * 
     * @type {UserUpdateDetails}
     * @memberof UserApiUpdateCurrentUser
     */
    readonly userUpdateDetails: UserUpdateDetails
}

/**
 * Request parameters for uploadUserLogo operation in UserApi.
 * @export
 * @interface UserApiUploadUserLogoRequest
 */
export interface UserApiUploadUserLogoRequest {
    /**
     * The image file to be uploaded.
     * @type {File}
     * @memberof UserApiUploadUserLogo
     */
    readonly file?: File
}

/**
 * Request parameters for verifyEmailOtp operation in UserApi.
 * @export
 * @interface UserApiVerifyEmailOtpRequest
 */
export interface UserApiVerifyEmailOtpRequest {
    /**
     * 
     * @type {VerifyEmailOtpRequest}
     * @memberof UserApiVerifyEmailOtp
     */
    readonly verifyEmailOtpRequest: VerifyEmailOtpRequest
}

/**
 * UserApi - object-oriented interface
 * @export
 * @class UserApi
 * @extends {BaseAPI}
 */
export class UserApi extends BaseAPI {
    /**
     * Returns information about the authenticated user, including their roles and permissions.
     * @summary Get current user information
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public getCurrentUser(options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).getCurrentUser(options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Logs in a user using credentials for a specific provider. Returns access and refresh tokens.
     * @summary Login with credentials from an authentication provider
     * @param {UserApiLoginRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public login(requestParameters: UserApiLoginRequest, options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).login(requestParameters.loginRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Registers a user with data provided from an authentication provider. If the user doesn\'t exist, creates a new account.
     * @summary Register a new user
     * @param {UserApiRegisterUserRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public registerUser(requestParameters: UserApiRegisterUserRequest, options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).registerUser(requestParameters.registerRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Resends a new OTP to the user\'s email address for verification.
     * @summary Resend OTP to user\'s email
     * @param {UserApiResendEmailOtpRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public resendEmailOtp(requestParameters: UserApiResendEmailOtpRequest, options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).resendEmailOtp(requestParameters.resendOtpRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Sends a one-time password (OTP) to the user\'s email address for verification purposes.
     * @summary Send OTP to email
     * @param {UserApiSendEmailOtpRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public sendEmailOtp(requestParameters: UserApiSendEmailOtpRequest, options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).sendEmailOtp(requestParameters.sendEmailOtpRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Allows the authenticated user to update their profile information, excluding the email.
     * @summary Update current user information
     * @param {UserApiUpdateCurrentUserRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public updateCurrentUser(requestParameters: UserApiUpdateCurrentUserRequest, options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).updateCurrentUser(requestParameters.userUpdateDetails, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Accepts an image file and responds with a URL to access the uploaded logo.
     * @summary Upload a user logo
     * @param {UserApiUploadUserLogoRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public uploadUserLogo(requestParameters: UserApiUploadUserLogoRequest = {}, options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).uploadUserLogo(requestParameters.file, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Verifies the one-time password (OTP) submitted by the user for email confirmation.
     * @summary Verify OTP for email
     * @param {UserApiVerifyEmailOtpRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof UserApi
     */
    public verifyEmailOtp(requestParameters: UserApiVerifyEmailOtpRequest, options?: RawAxiosRequestConfig) {
        return UserApiFp(this.configuration).verifyEmailOtp(requestParameters.verifyEmailOtpRequest, options).then((request) => request(this.axios, this.basePath));
    }
}